pyproject.toml
[project]
name = "doc-to-reword"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "pdfplumber>=0.10.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
